requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.7.0
pymongo==4.10.1
pymongo==4.10.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
